Using reference vmlinuz file for BIOS boot
The content of original vmlinuz, between the end of zeropage and
the start of kernel body, will be wiped out in the packing phase
of vbutil_kernel. With out this part, SeaBIOS or other real bios
systems can not boot the extracted kernel image.

This patch will take a reference vmlinuz file, inject the bios booting
part into the newly unpacked kernel, so the it will boot on BIOS
machines.